fix(rmm-eac5): WIP: Fix MISRA C defects
Fixes MISRA C 2012 Rule 2.3, 2.4, 8.3, 8.10,
10.3, 10.6, 12.1, 16.4, 18.3, 18.4, 21.2 and
21.15 violations.

+++ runtime/rmi/realm.c
@@ -125,7 +125,7 @@ static unsigned int s2_num_root_rtts(unsigned int ipa_bits, int sl)
 		return 1U;
 	}
 
-	return (1U << (ipa_bits - sl_ipa_bits));
+	return (unsigned int)(1UL << (ipa_bits - sl_ipa_bits));
 }
 
 /*
